Everything seems to build fine now!
If you're having trouble, try this:

Open the VxWorks Development Shell from the Start Menu and move to the directory where you have the project. Then run these commands:

rm CompetitionCode/PPC603gnu/Makefile
wrws_update.bat -data "$(repository directory)" --generate-makefiles
wrws_update.bat -data "$(repository directory)" --build-projects

but with the $(repository direct) replaced with where you have your repository. This destroys the makefile, which is the script that builds the project, and then regenerates it. After that, you can run the last command, which builds all of the projects. This should clean everything up, and after I did that everything built fine. I'll also see if I can get a script to listen on the VM in the background, so that way you can just tell it over the network to build and not have to worry about WindRiver as much.