Code for gyro/aim biases, updated dashboard feedback
The gyro bias is so that we can reset the gyro when its turned a certain
angle, like starting in autonomous.
The aim bias is implemented by first holding the robot angle (with or
without vision), and pressing either direction on the dpad to turn in
increments of 10 degrees to assist in aiming.